requirements.txt
streamlit
torch
torchvision
torchcam
Pillow
gdown